refactor(backup): extract import staging into backup/stage.rs
